separe functions in diferent files
