+++ python_aef/aef_publish.py
@@ -1,4 +1,3 @@
-
 import requests
 import json
 import configparser
@@ -8,44 +7,6 @@ import redis
 # Get environment variables
 REDIS_HOST = os.getenv('REDIS_HOST')
 REDIS_PORT = os.environ.get('REDIS_PORT')
-
-
-def register_exposer_to_capif(capif_ip, capif_port, username, password, role, description, cn):
-
-    print("Registering exposer to CAPIF")
-    url = "http://{}:{}/register".format(capif_ip, capif_port)
-
-    payload = dict()
-    payload['username'] = username
-    payload['password'] = password
-    payload['role'] = role
-    payload['description'] = description
-    payload['cn'] = cn
-
-    headers = {
-        'Content-Type': 'application/json'
-    }
-
-    try:
-        print("''''''''''REQUEST'''''''''''''''''")
-        print("Request: to ",url) 
-        print("Request Headers: ",  headers) 
-        print("Request Body: ", json.dumps(payload))
-        print("''''''''''REQUEST'''''''''''''''''")
-        response = requests.request("POST", url, headers=headers, data=json.dumps(payload))
-        response.raise_for_status()
-        response_payload = json.loads(response.text)
-        print("''''''''''RESPONSE'''''''''''''''''")
-        print("Response to: ",response.url) 
-        print("Response Headers: ",  response.headers) 
-        print("Response: ", response.json())
-        print("Response Status code: ", response.status_code)
-        print("Success to register new exposer")
-        print("''''''''''RESPONSE'''''''''''''''''")
-        return response_payload['id'], response_payload['ccf_publish_url']
-    except requests.exceptions.HTTPError as err:
-        raise Exception(err.response.text, err.response.status_code)
-
 
 def get_capif_auth(capif_ip, capif_port, username, password, role):
 
@@ -90,42 +51,6 @@ def get_capif_auth(capif_ip, capif_port, username, password, role):
     except requests.exceptions.HTTPError as err:
         raise Exception(err.response.text, err.response.status_code)
 
-def register_api_provider_to_capif(capif_ip, ccf_url):
-
-    print("Registering api provider to CAPIF")
-
-    url = 'https://{}/{}'.format(capif_ip, ccf_url)
-    payload = open('api_provider_domain.json', 'rb')
-    headers = {
-        'Content-Type': 'application/json'
-    }
-
-    try:
-
-        print("''''''''''REQUEST'''''''''''''''''")
-        print("Request: to ",url) 
-        print("Request Headers: ",  headers) 
-        print("Request Body: ", json.dumps(payload))
-        print("''''''''''REQUEST'''''''''''''''''")
-
-        response = requests.request("POST", url, headers=headers, data=payload, cert=('exposer.crt', 'private.key'), verify='ca.crt')
-        response.raise_for_status()
-        response_payload = json.loads(response.text)
-
-        print("''''''''''RESPONSE'''''''''''''''''")
-        print("Response to: ",response.url) 
-        print("Response Headers: ",  response.headers) 
-        print("Response: ", response.json())
-        print("Response Status code: ", response.status_code)
-        print("Success, registered api provider domain to CAPIF")
-        print("''''''''''RESPONSE'''''''''''''''''")
-        return response_payload['apiId']
-    except requests.exceptions.HTTPError as err:
-        message = json.loads(err.response.text)
-        status = err.response.status_code
-        raise Exception(message, status)
-
-
 def publish_service_api_to_capif(capif_ip, ccf_url):
 
     print("Publishing api service to CAPIF")
@@ -160,7 +85,6 @@ def publish_service_api_to_capif(capif_ip, ccf_url):
         status = err.response.status_code
         raise Exception(message, status)
 
-
 if __name__ == '__main__':
 
     r = redis.Redis(
@@ -182,35 +106,7 @@ if __name__ == '__main__':
     # capif_port = config.get("credentials", "capif_port")
     #capif_ip = os.getenv('CAPIF_HOSTNAME')
     capif_port = os.getenv('CAPIF_PORT')
-    
-    #First we need register exposer in CAPIF
-    try:
-        if not r.exists('exposerID'):
-            exposerID, ccf_publish_url = register_exposer_to_capif(capif_ip, capif_port, username, password, role, description,cn)
-            r.set('exposerID', exposerID)
-            r.set('ccf_publish_url', ccf_publish_url)
-            print("exposer ID: {}".format(exposerID))
-    except Exception as e:
-        status_code = e.args[0]
-        if status_code == 409:
-            print("User already registed. Continue with token request\n")
-        else:
-            print(e)
-
-    #Second, we need get auth, in this case create cert and private key file
-    try:
-        get_capif_auth(capif_ip, capif_port, username, password, role)
-
-    except Exception as e:
-        status_code = e.args[0]
-        if status_code == 401:
-            print("Bad credential or User not found\n")
-        else:
-            print(e)
-        capif_access_token = None
-
-
-    #Third publish service in CAPIF
+#Third publish service in CAPIF
     try:
         if r.exists('exposerID'):
             ccf_publish_url = r.get('ccf_publish_url')
